Moved OpenGL context to GraphicsDevice.OpenGL
MonoGame is not using any of the higher-level facilities of
OpenTK.GameWindow. Conceptually, it makes sense to base MonoGame’s
OpenTKGameWindow on OpenTK.NativeWindow and move the OpenGL context to
GraphicsDevice.OpenGL.

This has two advantages:
1. MonoGame applications no longer initialize OpenGL unnecessarily
before instantiating a GraphicsDevice service.
2. GraphicsDevice.OpenGL has enough information to setup the correct
OpenGL attributes for color, depth and stencil depth, as well as
multisampling. This was not possible before, since the
PresentationParameters are not available when a GameWindow is first
constructed.